rm l2_normalize_util
