+++ oneflow/core/job_completer/square_grad.cpp
@@ -16,7 +16,6 @@ void GenerateBackwardOpConf(
     multiply_in_op_conf->set_in_1(GenLogicalBlobName(op.BnInOp2Lbi("in")));
     multiply_in_op_conf->set_out("out");
     op_confs->push_back(multiply_in_op);
-
     OperatorConf scalar_mul_op;
     scalar_mul_op.set_name(op.op_name() + "_scalar_mul_in_grad");
     ScalarMulOpConf* scalar_mul_op_conf = scalar_mul_op.mutable_scalar_mul_conf();
@@ -24,9 +23,8 @@ void GenerateBackwardOpConf(
     scalar_mul_op_conf->set_in(multiply_in_op.name() + "/out");
     scalar_mul_op_conf->set_out("out");
     op_confs->push_back(scalar_mul_op);
-
     DiffLbi4BnInOp("in")->set_op_name(scalar_mul_op.name());
-    DiffLbi4BnInOp("in")->set_blob_name("out");
+    DiffLbi4BnInOp("in")->set_blob_name(scalar_mul_op_conf->out());
   }
 }
