+++ oneflow/core/job_completer/where_grad.cpp
@@ -12,7 +12,7 @@ void GenerateBackwardOpConf(
   zeros_like_op.set_name(op.op_name() + "_zeros");
   ZerosLikeOpConf* zeros_like_op_conf = zeros_like_op.mutable_zeros_like_conf();
   zeros_like_op_conf->set_like(GenLogicalBlobName(op.BnInOp2Lbi("x")));
-  zeros_like_op_conf->set_y("y");
+  zeros_like_op_conf->set_out("out");
   op_confs->push_back(zeros_like_op);
 
   if (DiffLbi4BnInOp("x") != nullptr) {
@@ -21,7 +21,7 @@ void GenerateBackwardOpConf(
     WhereOpConf* where_x_grad_op_conf = where_x_grad_op.mutable_where_conf();
     where_x_grad_op_conf->set_condition(GenLogicalBlobName(op.BnInOp2Lbi("condition")));
     where_x_grad_op_conf->set_x(GenLogicalBlobName(*DiffLbi4BnInOp("out")));
-    where_x_grad_op_conf->set_y(zeros_like_op.name() + "/y");
+    where_x_grad_op_conf->set_y(zeros_like_op.name() + "/out");
     where_x_grad_op_conf->set_out("out");
     op_confs->push_back(where_x_grad_op);
 
@@ -33,7 +33,7 @@ void GenerateBackwardOpConf(
     where_y_grad_op.set_name(op.op_name() + "_y_grad");
     WhereOpConf* where_y_grad_op_conf = where_y_grad_op.mutable_where_conf();
     where_y_grad_op_conf->set_condition(GenLogicalBlobName(op.BnInOp2Lbi("condition")));
-    where_y_grad_op_conf->set_x(zeros_like_op.name() + "/y");
+    where_y_grad_op_conf->set_x(zeros_like_op.name() + "/out");
     where_y_grad_op_conf->set_y(GenLogicalBlobName(*DiffLbi4BnInOp("out")));
     where_y_grad_op_conf->set_out("out");
     op_confs->push_back(where_y_grad_op);
