+++ oneflow/core/kernel/l2_normalize_kernel.cu
@@ -1,7 +1,6 @@
-#include "oneflow/core/kernel/l2_normalize_kernel_util.h"
-#include "oneflow/core/kernel/kernel_util.h"
+#include "oneflow/core/kernel/kernel.h"
 #include <cub/cub.cuh>
-#include <math.h>
+
 namespace oneflow {
 
 namespace {
@@ -69,34 +68,58 @@ __global__ void L2NormalizeBackward(const int32_t n, const int32_t c, const int3
 }  // namespace
 
 template<typename T>
-struct L2NormalizeKernelUtil<kGPU, T> {
-  static void Forward(DeviceCtx* ctx, const int32_t axis, const float epsilon, const Blob* in_blob,
-                      Blob* square_x_sum_blob, Blob* out_blob) {
+class L2NormalizeGpuKernel final : public KernelIf<DeviceType::kGPU> {
+ public:
+  OF_DISALLOW_COPY_AND_MOVE(L2NormalizeGpuKernel);
+  L2NormalizeGpuKernel() = default;
+  ~L2NormalizeGpuKernel() = default;
+
+ private:
+  void ForwardDataContent(const KernelCtx& ctx,
+                          std::function<Blob*(const std::string&)> BnInOp2Blob) const override {
+    const Blob* in_blob = BnInOp2Blob("in");
+    const int32_t axis = this->op_conf().l2_normalize_conf().axis();
+    const float epsilon = this->op_conf().l2_normalize_conf().epsilon();
     int32_t c = in_blob->shape().At(axis);
     int32_t n = in_blob->shape().elem_cnt() / c;
     int32_t d = in_blob->shape().Count(axis + 1);
     L2NormalizeForward<<<std::min(n, kCudaMaxBlocksNum), kCudaThreadsNumPerBlock, 0,
-                         ctx->cuda_stream()>>>(n, c, d, static_cast<T>(epsilon), in_blob->dptr<T>(),
-                                               square_x_sum_blob->mut_dptr<T>(),
-                                               out_blob->mut_dptr<T>());
-  }
-
-  static void Backward(DeviceCtx* ctx, const int32_t axis, const float epsilon,
-                       const Blob* out_blob, const Blob* out_diff_blob,
-                       const Blob* square_x_sum_blob, Blob* in_diff_blob) {
-    int32_t c = out_blob->shape().At(axis);
-    int32_t n = out_blob->shape().elem_cnt() / c;
-    int32_t d = out_blob->shape().Count(axis + 1);
-    L2NormalizeBackward<<<std::min(n, kCudaMaxBlocksNum), kCudaThreadsNumPerBlock, 0,
-                          ctx->cuda_stream()>>>(
-        n, c, d, static_cast<T>(epsilon), out_blob->dptr<T>(), out_diff_blob->dptr<T>(),
-        square_x_sum_blob->dptr<T>(), in_diff_blob->mut_dptr<T>());
+                         ctx.device_ctx->cuda_stream()>>>(
+        n, c, d, static_cast<T>(epsilon), in_blob->dptr<T>(),
+        BnInOp2Blob("square_x_sum")->mut_dptr<T>(), BnInOp2Blob("out")->mut_dptr<T>());
   }
 };
 
-#define INSTANTIATE_L2_NORMALIZE_KERNEL_UTIL_GPU(type_cpp, type_proto) \
-  template struct L2NormalizeKernelUtil<DeviceType::kGPU, type_cpp>;
-OF_PP_FOR_EACH_TUPLE(INSTANTIATE_L2_NORMALIZE_KERNEL_UTIL_GPU, ARITHMETIC_DATA_TYPE_SEQ)
-#undef INSTANTIATE_L2_NORMALIZE_KERNEL_UTIL_GPU
+template<typename T>
+class L2NormalizeGradGpuKernel final : public KernelIf<DeviceType::kGPU> {
+ public:
+  OF_DISALLOW_COPY_AND_MOVE(L2NormalizeGradGpuKernel);
+  L2NormalizeGradGpuKernel() = default;
+  ~L2NormalizeGradGpuKernel() = default;
+
+ private:
+  void ForwardDataContent(const KernelCtx& ctx,
+                          std::function<Blob*(const std::string&)> BnInOp2Blob) const override {
+    const Blob* dy_blob = BnInOp2Blob("dy");
+    const int32_t axis = this->op_conf().l2_normalize_grad_conf().axis();
+    const float epsilon = this->op_conf().l2_normalize_grad_conf().epsilon();
+    int32_t c = dy_blob->shape().At(axis);
+    int32_t n = dy_blob->shape().elem_cnt() / c;
+    int32_t d = dy_blob->shape().Count(axis + 1);
+    L2NormalizeBackward<<<std::min(n, kCudaMaxBlocksNum), kCudaThreadsNumPerBlock, 0,
+                          ctx.device_ctx->cuda_stream()>>>(
+        n, c, d, static_cast<T>(epsilon), BnInOp2Blob("y")->dptr<T>(), dy_blob->dptr<T>(),
+        BnInOp2Blob("square_x_sum")->dptr<T>(), BnInOp2Blob("dx")->mut_dptr<T>());
+  }
+};
+
+REGISTER_KERNEL_WITH_DEVICE_AND_DTYPE(OperatorConf::kL2NormalizeConf, DeviceType::kGPU, float,
+                                      L2NormalizeGpuKernel<float>)
+REGISTER_KERNEL_WITH_DEVICE_AND_DTYPE(OperatorConf::kL2NormalizeConf, DeviceType::kGPU, double,
+                                      L2NormalizeGpuKernel<double>)
+REGISTER_KERNEL_WITH_DEVICE_AND_DTYPE(OperatorConf::kL2NormalizeGradConf, DeviceType::kGPU, float,
+                                      L2NormalizeGradGpuKernel<float>)
+REGISTER_KERNEL_WITH_DEVICE_AND_DTYPE(OperatorConf::kL2NormalizeGradConf, DeviceType::kGPU, double,
+                                      L2NormalizeGradGpuKernel<double>)
 
 }  // namespace oneflow
