+++ oneflow/core/operator/zeros_like_op.cpp
@@ -11,7 +11,7 @@ class ZerosLikeOp final : public Operator {
 
   void InitFromOpConf() override {
     CHECK(op_conf().has_zeros_like_conf());
-    EnrollOutputBn("y");
+    EnrollOutputBn("out");
     EnrollInputBn("like", false)->set_use_header_only(true);
   }
   const PbMessage& GetCustomizedConf() const override { return op_conf().zeros_like_conf(); }
@@ -19,12 +19,12 @@ class ZerosLikeOp final : public Operator {
   Maybe<void> InferBlobDescs(std::function<BlobDesc*(const std::string&)> GetBlobDesc4BnInOp,
                              const ParallelContext* parallel_ctx) const override {
     const BlobDesc* like_blob_desc = GetBlobDesc4BnInOp("like");
-    BlobDesc* y_blob_desc = GetBlobDesc4BnInOp("y");
-    GetBlobDesc4BnInOp("y")->CopyMetaFrom(*GetBlobDesc4BnInOp("like"));
+    BlobDesc* out_blob_desc = GetBlobDesc4BnInOp("out");
+    GetBlobDesc4BnInOp("out")->CopyMetaFrom(*GetBlobDesc4BnInOp("like"));
     if (op_conf().zeros_like_conf().data_type()) {
-      y_blob_desc->set_data_type(op_conf().zeros_like_conf().data_type());
+      out_blob_desc->set_data_type(op_conf().zeros_like_conf().data_type());
     } else {
-      y_blob_desc->set_data_type(like_blob_desc->data_type());
+      out_blob_desc->set_data_type(like_blob_desc->data_type());
     }
     return Maybe<void>::Ok();
   }
